src/api-NLP/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
asyncio==3.4.3
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confection==0.1.3
cymem==2.0.7
Flask==2.3.3
Flask-Cors==4.0.0
gensim==4.3.2
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pathy==0.10.2
pluggy==1.3.0
preshed==3.0.8
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.6.0/pt_core_news_lg-3.6.0-py3-none-any.whl#sha256=4a15fab02cd32ba7bc2256c73ad453eb9d307f1393d825d77cb296028422deb3
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==7.4.2
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.2
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.12
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
websocket-client==1.6.1
Werkzeug==2.3.7
zipp==3.16.2
